Lab Assignment # 9
COP 3502
Lab Assignment # 9
In this lab, you are going to implement hashing with chaining with separate chaining. Please implement
your hash function accordingly to distribute the data evenly in the array.
1. You are going to implement the hash function bodies.
2. Your submission will be the repository url of your implementation. Per your submission, the TAs
will be able to clone your repository and compile and run it.
3. If you have specific build instructions, please update your README file accordingly.
4. You cannot make any updates to your code after the submission deadline. Even if you do so, we
will be able to track your updates, hence you might lose points for making a change after the
submission deadline.